fixed css styles for button, check for user is null removed before the main return section, created enum, moved types to  separate file, fixed filter method using
