+++ src/components/CurrentUser/CurrentUser.tsx
@@ -8,44 +8,50 @@ export const CurrentUser: React.FC = () => {
   const dispatch = useDispatch();
   const user = useSelector(getUserSelector);
 
+  if (user === null) {
+    return (
+      <div>
+        No user selected!!!
+      </div>
+    );
+  }
+
   return (
     <div className="CurrentUser">
-      {user === null ? ('No user selected') : (
-        <>
-          {user.id === 0 ? (
-            'Error, user not found!!!'
-          ) : (
-            <>
-              <h2 className="CurrentUser__title">
-                <span>
-                  {`Selected user: ${user.id}`}
-                </span>
-              </h2>
+      <>
+        {user.id === 0 ? (
+          'Error, user not found!!!'
+        ) : (
+          <>
+            <h2 className="CurrentUser__title">
+              <span>
+                {`Selected user: ${user.id}`}
+              </span>
+            </h2>
 
-              <h3
-                className="CurrentUser__name"
-                data-cy="userName"
+            <h3
+              className="CurrentUser__name"
+              data-cy="userName"
+            >
+              {user.name}
+            </h3>
+            <p className="CurrentUser__email">{user.email}</p>
+            <p className="CurrentUser__phone">{user.phone}</p>
+            <h3>
+              <button
+                type="button"
+                className="CurrentUser__title button"
+                onClick={() => {
+                  dispatch(setUserAction(null));
+                  dispatch(setSelectedTodoId(0));
+                }}
               >
-                {user.name}
-              </h3>
-              <p className="CurrentUser__email">{user.email}</p>
-              <p className="CurrentUser__phone">{user.phone}</p>
-              <h3>
-                <button
-                  type="button"
-                  className="CurrentUser__title button"
-                  onClick={() => {
-                    dispatch(setUserAction(null));
-                    dispatch(setSelectedTodoId(0));
-                  }}
-                >
-                  Clear
-                </button>
-              </h3>
-            </>
-          )}
-        </>
-      )}
+                Clear
+              </button>
+            </h3>
+          </>
+        )}
+      </>
     </div>
   );
 };
