+++ src/components/TodoList/TodoList.scss
@@ -10,7 +10,7 @@
   &__user-button {
     width: 80px;
     height: 30px;
-    font-size: 14px;
+    font-size: 11px;
 
     &--remove {
       margin-left: 5px;
